requirements.txt
openai==2.8.1
python-dotenv==1.2.1
streamlit==1.51.0